+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Fix linking of identifiers that contain unicode characters.
   ([#831](https://github.com/mitmproxy/pdoc/issues/831), @iFreilicht)
+- Add support for Pydantic-style field docstrings,
+  e.g. `pydantic.Field(description="...")`
 
 ## 2025-06-04: pdoc 15.0.4