+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased: pdoc next
 
+- pdoc has a new logo. 🐍
+  ([#838](https://github.com/mitmproxy/pdoc/pull/838), @mhils)
 - Fix linking of identifiers that contain unicode characters.
   ([#831](https://github.com/mitmproxy/pdoc/issues/831), @iFreilicht)
 - Replace vendored version of `markdown2` with the [official